requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
# --------- pytorch --------- #
timm==0.6.7
torch==1.11.0+cpu 
torchvision==0.12.0+cpu
pytorch-lightning==1.7.1
torchmetrics==0.9.3

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
# hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
# pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
# pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)

# --------- deployment --------- #
gradio==3.3.1

# --------- explainability --------- #
captum 
shap 
grad-cam
